RDDID Estimation

This function estimates the ATU or ATT using a RDDID framework, assuming potential outcome discontinuities are constant across time.

Usage

mult_period_est(
  df,
  t_star,
  t_vec,
  w_vec,
  h,
  b,
  tname = "time",
  xname = "R",
  yname = "Y",
  idname = "id"
)

Arguments

df

dataframe

t_star

time period with RD treatment

t_vec

time period for bias estimation

w_vec

weights for t_vec time periods

h

bandwidth

b

bias correction bandwidth

tname

the name of the time period variable in df

xname

the name of the running variable in df

yname

the name of the outcome variable in df

idname

the name of the id variable in df

Value

Conventional (C) and Bias Corrected (BC) treatment effect estimates and standard errors (SE). SEs are provided for sampling under repeated cross section (CS), panel with constant running variable (PC) and varying running variable (PV).